requirements.txt
pandas>=1.4.0
openpyxl>=3.0.0
openai>=0.27.0
PyYAML>=6.0
python-dotenv>=0.19.0
